requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
drf-nested-routers==0.93.4
email-validator==2.1.0.post1
googlemaps==4.10.0
idna==3.4
Pillow==10.1.0
psycopg2-binary==2.9.3
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.6
